backend/requirements.txt
Flask==2.0.2
requests==2.26.0
psycopg2-binary==2.9.1
Jinja2==3.0.0
MarkupSafe==2.0.1

